# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quantum FX Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trade/execute:
    post:
      operationId: executeTrade
      tags: [trade]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteTradeInput"
      responses:
        "200":
          description: Opened trade position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradePosition"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trade/positions:
    get:
      operationId: listTradePositions
      tags: [trade]
      responses:
        "200":
          description: User trade positions (open and closed)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradePosition"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trade/positions/{id}/close:
    post:
      operationId: closeTradePosition
      tags: [trade]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed trade position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradePosition"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Team
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ExecuteTradeInput:
      type: object
      properties:
        signalId:
          type: string
        botId:
          type: integer
        targetProfit:
          type: number
        stopLoss:
          type: number
        stake:
          type: number
      required: [signalId, botId, targetProfit, stopLoss, stake]
...
    TradePosition:
      type: object
      properties:
        id:
          type: integer
        signalId:
          type: string
        pair:
          type: string
        direction:
          type: string
        market:
          type: string
        botId:
          type: integer
        botName:
          type: string
        stake:
          type: number
        targetProfit:
          type: number
        stopLoss:
          type: number
        status:
          type: string
        pnl:
          type: number
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
        elapsedMs:
          type: integer
      required: [id, signalId, pair, direction, market, botId, botName, stake, targetProfit, stopLoss, status, pnl, openedAt, closedAt, elapsedMs]

    # Auth
...